encapsulate AO creation, create client registry
This is a bit cludgy, but having a private (truly private) derived class that produces the static 'create()' method allows for the base class to have a protected constructor (in theory, not going to try to finagle that yet) and it keeps the base class from having to declare a static create() function in its class definition.

Of course that means the ->m_hp= hack no longer works, and have to properly (read: tons of boilerplate) initialize it in the constructor.

Also, I suppose you could use dynamic_cast to tell whether a GenericCAO was constructed outside of the registry or not, FWIW