instrument main() to setup registries
And so, to keep this from being in nebulous global initialization, I actually have to find the places in main to call the setup functions. Non-trivial, since the clientRegistry shouldn't exist for pure servers (if SERVER is defined)

Also more cruft fixing thanks to compile errors.

Also why did I get rid of that getType() abstraction in the default getSendType()...